Extract xpath from html transform
